+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ### 1.0.1 - 14-Jul-2014
 
 - Bugfix for PHP fatal error when using the trait.
+- Added optional `$castOnSet` property to cast attributes when setting as well.
 
 
 ### 1.0.0 - 10-Jul-2014